+++ src/configs/proxy-config.service.ts
@@ -17,6 +17,7 @@ export class ProxyConfigService implements ProxyModuleOptionsFactory {
       {
         id: 'jsonplaceholder',
         url: 'https://jsonplaceholder.typicode.com',
+        forwardToken: false,
       },
     ];